pyproject.toml
[project]
name = "django-unfold-modal"
version = "0.1.0"
description = "Modal-based related-object popups for django-unfold"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "django>=5.0",
    "django-unfold>=0.52.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Repository = "https://github.com/metaforx/django-unfold-modal"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["unfold_modal"]

# Poetry for dependency management (dev/test)
[tool.poetry]
name = "django-unfold-modal"
version = "0.1.0"
description = "Modal-based related-object popups for django-unfold"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
django = ">=5.0"
django-unfold = ">=0.52.0"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0"
pytest-django = ">=4.5"
pytest-playwright = ">=0.4"
pytest-env = "^1.2.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.4"

[tool.pytest.ini_options]
addopts = """\
    --strict-config
    --strict-markers
    --ds=testapp.settings
    """
testpaths = ["tests"]
pythonpath = ["tests/server"]
django_find_project = false
env = [
    "DJANGO_ALLOW_ASYNC_UNSAFE=true",
]

[tool.ruff]
fix = true
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B905", # zip() without strict=True
    "C901", # too complex
]
